+++ library/src/test/java/io/github/luizgrp/sectionedrecyclerviewadapter/SectionTest.java
@@ -12,6 +12,7 @@ import io.github.luizgrp.sectionedrecyclerviewadapter.testdoubles.spy.BindingHea
 import io.github.luizgrp.sectionedrecyclerviewadapter.testdoubles.spy.BindingSectionSpy;
 import io.github.luizgrp.sectionedrecyclerviewadapter.testdoubles.stub.SectionStub;
 
+import static io.github.luizgrp.sectionedrecyclerviewadapter.Section.State;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
@@ -31,7 +32,108 @@ public class SectionTest {
     }
 
     @Test
-    public void onBindViewHolder_Section_isCalled() {
+    public void constructor_withSectionParameters_constructsCorrectSection() {
+        // Given
+        final int ITEM_ID = 1;
+        final int HEADER_ID = 2;
+        final int FOOTER_ID = 3;
+        final int FAILED_ID = 4;
+        final int LOADING_ID = 5;
+        final int EMPTY_ID = 6;
+
+        @SuppressWarnings("ResourceType")
+        SectionParameters sectionParameters = new SectionParameters.Builder(ITEM_ID)
+                .headerResourceId(HEADER_ID)
+                .footerResourceId(FOOTER_ID)
+                .failedResourceId(FAILED_ID)
+                .loadingResourceId(LOADING_ID)
+                .emptyResourceId(EMPTY_ID)
+                .build();
+        Section section = getSection(sectionParameters);
+
+        // When
+        int resultItemId = section.getItemResourceId();
+        int resultHeaderId = section.getHeaderResourceId();
+        int resultFooterId = section.getFooterResourceId();
+        int resultFailedId = section.getFailedResourceId();
+        int resultLoadingId = section.getLoadingResourceId();
+        int resultEmptyId = section.getEmptyResourceId();
+        boolean resultHasHeader = section.hasHeader();
+        boolean resultHasFooter = section.hasFooter();
+
+        // Then
+        assertThat(resultItemId, is(ITEM_ID));
+        assertThat(resultHeaderId, is(HEADER_ID));
+        assertThat(resultFooterId, is(FOOTER_ID));
+        assertThat(resultFailedId, is(FAILED_ID));
+        assertThat(resultLoadingId, is(LOADING_ID));
+        assertThat(resultEmptyId, is(EMPTY_ID));
+        assertThat(resultHasHeader, is(true));
+        assertThat(resultHasFooter, is(true));
+    }
+
+    public void setState_withValidLoadingResId_succeeds() {
+        // Given
+        final int ITEM_ID = 1;
+        final int LOADING_ID = 2;
+
+        @SuppressWarnings("ResourceType")
+        SectionParameters sectionParameters = new SectionParameters.Builder(ITEM_ID)
+                .loadingResourceId(LOADING_ID)
+                .build();
+        Section section = getSection(sectionParameters);
+
+        // When
+        section.setState(State.LOADING);
+
+        // Then
+        assertThat(section.getState(), is(State.LOADING));
+    }
+    
+    @Test(expected = IllegalStateException.class)
+    public void setState_withMissingLoadingResId_throwsException() {
+        // Given
+        final int ITEM_ID = 1;
+        
+        @SuppressWarnings("ResourceType")
+        SectionParameters sectionParameters = new SectionParameters.Builder(ITEM_ID)
+                .build();
+        Section section = getSection(sectionParameters);
+
+        // When
+        section.setState(State.LOADING);
+    }
+
+    @Test(expected = IllegalStateException.class)
+    public void setState_withMissingFailedResId_throwsException() {
+        // Given
+        final int ITEM_ID = 1;
+
+        @SuppressWarnings("ResourceType")
+        SectionParameters sectionParameters = new SectionParameters.Builder(ITEM_ID)
+                .build();
+        Section section = getSection(sectionParameters);
+
+        // When
+        section.setState(State.FAILED);
+    }
+
+    @Test(expected = IllegalStateException.class)
+    public void setState_withEmptyFailedResId_throwsException() {
+        // Given
+        final int ITEM_ID = 1;
+
+        @SuppressWarnings("ResourceType")
+        SectionParameters sectionParameters = new SectionParameters.Builder(ITEM_ID)
+                .build();
+        Section section = getSection(sectionParameters);
+
+        // When
+        section.setState(State.EMPTY);
+    }
+
+    @Test
+    public void onBindViewHolder_withSection_isCalled() {
         // Given
         sectionAdapter.addSection(new SectionStub(ITEMS_QTY));
 
@@ -47,14 +149,14 @@ public class SectionTest {
     }
 
     @Test
-    public void onBindLoadingViewHolder_Section_isCalled() {
+    public void onBindLoadingViewHolder_withSection_isCalled() {
         // Given
         sectionAdapter.addSection(new SectionStub(ITEMS_QTY));
 
         BindingSectionSpy sectionSpy = new BindingSectionSpy(ITEMS_QTY);
         sectionAdapter.addSection(sectionSpy);
 
-        sectionSpy.setState(Section.State.LOADING);
+        sectionSpy.setState(State.LOADING);
 
         // When
         // Section - Loading [10]
@@ -65,14 +167,14 @@ public class SectionTest {
     }
 
     @Test
-    public void onBindFailedViewHolder_Section_isCalled() {
+    public void onBindFailedViewHolder_withSection_isCalled() {
         // Given
         sectionAdapter.addSection(new SectionStub(ITEMS_QTY));
 
         BindingSectionSpy sectionSpy = new BindingSectionSpy(ITEMS_QTY);
         sectionAdapter.addSection(sectionSpy);
 
-        sectionSpy.setState(Section.State.FAILED);
+        sectionSpy.setState(State.FAILED);
 
         // When
         // Section - Failed [10]
@@ -83,7 +185,25 @@ public class SectionTest {
     }
 
     @Test
-    public void onBindViewHolder_HeadedSection_isCalled() {
+    public void onBindEmptyViewHolder_withSection_isCalled() {
+        // Given
+        sectionAdapter.addSection(new SectionStub(ITEMS_QTY));
+
+        BindingSectionSpy sectionSpy = new BindingSectionSpy(ITEMS_QTY);
+        sectionAdapter.addSection(sectionSpy);
+
+        sectionSpy.setState(State.EMPTY);
+
+        // When
+        // Section - Empty [10]
+        sectionAdapter.onBindViewHolder(null, 10);
+
+        // Then
+        assertTrue(sectionSpy.onBindEmptyViewHolderWasCalled);
+    }
+
+    @Test
+    public void onBindViewHolder_withHeadedSection_isCalled() {
         // Given
         sectionAdapter.addSection(new SectionStub(ITEMS_QTY));
 
@@ -102,7 +222,7 @@ public class SectionTest {
     }
 
     @Test
-    public void onBindViewHolder_FootedSection_isCalled() {
+    public void onBindViewHolder_withFootedSection_isCalled() {
         // Given
         sectionAdapter.addSection(new SectionStub(ITEMS_QTY));
 
@@ -121,7 +241,7 @@ public class SectionTest {
     }
 
     @Test
-    public void onBindViewHolder_HeadedFootedSection_isCalled() {
+    public void onBindViewHolder_withHeadedFootedSection_isCalled() {
         // Given
         sectionAdapter.addSection(new SectionStub(ITEMS_QTY));
 
@@ -140,43 +260,6 @@ public class SectionTest {
         assertTrue(sectionSpy.onBindHeaderViewHolderWasCalled);
         assertTrue(sectionSpy.onBindItemViewHolderWasCalled);
         assertTrue(sectionSpy.onBindFooterViewHolderWasCalled);
-    }
-
-    @Test
-    public void build_constructsCorrectSection() {
-        // Given
-        final int ITEM_ID = 1;
-        final int HEADER_ID = 2;
-        final int FOOTER_ID = 3;
-        final int FAILED_ID = 4;
-        final int LOADING_ID = 5;
-
-        @SuppressWarnings("ResourceType")
-        SectionParameters sectionParameters = new SectionParameters.Builder(ITEM_ID)
-                .headerResourceId(HEADER_ID)
-                .footerResourceId(FOOTER_ID)
-                .failedResourceId(FAILED_ID)
-                .loadingResourceId(LOADING_ID)
-                .build();
-        Section section = getSection(sectionParameters);
-
-        // When
-        int resultItemId = section.getItemResourceId();
-        int resultHeaderId = section.getHeaderResourceId();
-        int resultFooterId = section.getFooterResourceId();
-        int resultFailedId = section.getFailedResourceId();
-        int resultLoadingId = section.getLoadingResourceId();
-        boolean resultHasHeader = section.hasHeader();
-        boolean resultHasFooter = section.hasFooter();
-
-        // Then
-        assertThat(resultItemId, is(ITEM_ID));
-        assertThat(resultHeaderId, is(HEADER_ID));
-        assertThat(resultFooterId, is(FOOTER_ID));
-        assertThat(resultFailedId, is(FAILED_ID));
-        assertThat(resultLoadingId, is(LOADING_ID));
-        assertThat(resultHasHeader, is(true));
-        assertThat(resultHasFooter, is(true));
     }
 
     private Section getSection(SectionParameters sectionParameters) {