+++ library/src/test/java/io/github/luizgrp/sectionedrecyclerviewadapter/testdoubles/spy/BindingHeadedFootedSectionSpy.java
@@ -16,6 +16,7 @@ public class BindingHeadedFootedSectionSpy extends Section {
     public boolean onBindFooterViewHolderWasCalled = false;
     public boolean onBindLoadingViewHolderWasCalled = false;
     public boolean onBindFailedViewHolderWasCalled = false;
+    public boolean onBindEmptyViewHolderWasCalled = false;
 
     private final int contentItemsTotal;
 
@@ -25,6 +26,7 @@ public class BindingHeadedFootedSectionSpy extends Section {
                 .footerResourceId(-1)
                 .failedResourceId(-1)
                 .loadingResourceId(-1)
+                .emptyResourceId(-1)
                 .build());
 
         this.contentItemsTotal = contentItemsTotal;
@@ -63,5 +65,10 @@ public class BindingHeadedFootedSectionSpy extends Section {
     @Override
     public void onBindFailedViewHolder(RecyclerView.ViewHolder holder) {
         onBindFailedViewHolderWasCalled = true;
+    }
+
+    @Override
+    public void onBindEmptyViewHolder(RecyclerView.ViewHolder holder) {
+        onBindEmptyViewHolderWasCalled = true;
     }
 }